requirements.txt
appdirs==1.4.3
APScheduler==3.9.1
apturl==0.5.2
azure-core==1.26.0
azure-identity==1.11.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==0.10.1
Flask-Mail==0.9.1
Flask-WTF==0.11
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==0.24
Jinja2==2.7.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==0.23
monotonic==1.5
more-itertools==4.2.0
msal==1.20.0
msal-extensions==1.0.0
msgraph-core==0.2.2
netifaces==0.10.4
numpy==1.23.3
oauthlib==3.1.0
olefile==0.46
pandas==1.5.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
portalocker==2.5.1
protobuf==3.6.1
psycopg2-binary==2.9.4
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==6.0
remarkable==1.87
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
typing-extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
Werkzeug==0.10.1
WTForms==2.0.2
xkit==0.0.0
zipp==1.0.0
